separated job details into its own component
